+++ Realm/RLMResults.mm
@@ -33,6 +33,103 @@
 
 using namespace realm;
 
+static const int RLMEnumerationBufferSize = 16;
+
+@implementation RLMFastEnumerator {
+    // The buffer supplied by fast enumeration does not retain the objects given
+    // to it, but because we create objects on-demand and don't want them
+    // autoreleased (a table can have more rows than the device has memory for
+    // accessor objects) we need a thing to retain them.
+    id _strongBuffer[RLMEnumerationBufferSize];
+
+    RLMRealm *_realm;
+    RLMObjectSchema *_objectSchema;
+
+    // Collection being enumerated. Only one of these two will be valid: when
+    // possible we enumerate the collection directly, but when in a write
+    // transaction we instead create a frozen TableView and enumerate that
+    // instead so that mutating the collection during enumeration works.
+    id<RLMFastEnumerable> _collection;
+    realm::TableView _tableView;
+}
+
+- (instancetype)initWithCollection:(id<RLMFastEnumerable>)collection objectSchema:(RLMObjectSchema *)objectSchema {
+    self = [super init];
+    if (self) {
+        _realm = collection.realm;
+        _objectSchema = objectSchema;
+
+        if (_realm.inWriteTransaction) {
+            _tableView = [collection tableView];
+        }
+        else {
+            _collection = collection;
+            [_realm registerEnumerator:self];
+        }
+    }
+    return self;
+}
+
+- (void)dealloc {
+    if (_collection) {
+        [_realm unregisterEnumerator:self];
+    }
+}
+
+- (void)detach {
+    _tableView = [_collection tableView];
+    _collection = nil;
+}
+
+- (NSUInteger)countByEnumeratingWithState:(NSFastEnumerationState *)state
+                                    count:(NSUInteger)len {
+    RLMCheckThread(_realm);
+    if (!_tableView.is_attached() && !_collection) {
+        @throw RLMException(@"Collection is no longer valid");
+    }
+    // The fast enumeration buffer size is currently a hardcoded number in the
+    // compiler so this can't actually happen, but just in case it changes in
+    // the future...
+    if (len > RLMEnumerationBufferSize) {
+        len = RLMEnumerationBufferSize;
+    }
+
+    NSUInteger batchCount = 0, count = state->extra[1];
+
+    Class accessorClass = _objectSchema.accessorClass;
+    for (NSUInteger index = state->state; index < count && batchCount < len; ++index) {
+        size_t row = _collection ? [_collection indexInSource:index] : _tableView.get_source_ndx(index);
+
+        RLMObject *accessor = [[accessorClass alloc] initWithRealm:_realm schema:_objectSchema];
+        accessor->_row = (*_objectSchema.table)[row];
+        _strongBuffer[batchCount] = accessor;
+        batchCount++;
+    }
+
+    for (NSUInteger i = batchCount; i < len; ++i) {
+        _strongBuffer[i] = nil;
+    }
+
+    if (batchCount == 0) {
+        // Release our data if we're done, as we're autoreleased and so may
+        // stick around for a while
+        _collection = nil;
+        if (_tableView.is_attached()) {
+            _tableView = TableView();
+        }
+        else {
+            [_realm unregisterEnumerator:self];
+        }
+    }
+
+    state->itemsPtr = (__unsafe_unretained id *)(void *)_strongBuffer;
+    state->state += batchCount;
+    state->mutationsPtr = state->extra+1;
+
+    return batchCount;
+}
+@end
+
 //
 // RLMResults implementation
 //
@@ -136,47 +233,21 @@ static inline void RLMResultsValidateInWriteTransaction(__unsafe_unretained RLMR
 }
 
 - (NSUInteger)countByEnumeratingWithState:(NSFastEnumerationState *)state
-                                  objects:(__unsafe_unretained id [])buffer
+                                  objects:(__unused __unsafe_unretained id [])buffer
                                     count:(NSUInteger)len {
-    RLMResultsValidate(self);
-
-    __autoreleasing RLMCArrayHolder *items;
+    __autoreleasing RLMFastEnumerator *enumerator;
     if (state->state == 0) {
-        items = [[RLMCArrayHolder alloc] initWithSize:len];
-        state->extra[0] = (long)items;
+        RLMResultsValidate(self);
+
+        enumerator = [[RLMFastEnumerator alloc] initWithCollection:self objectSchema:_objectSchema];
+        state->extra[0] = (long)enumerator;
         state->extra[1] = self.count;
     }
     else {
-        // FIXME: mutationsPtr should be pointing to a value updated by core
-        // whenever the results are changed rather than doing this check
-        if (state->extra[1] != self.count) {
-            @throw RLMException(@"Collection was mutated while being enumerated.");
-        }
-        items = (__bridge id)(void *)state->extra[0];
-        [items resize:len];
+        enumerator = (__bridge id)(void *)state->extra[0];
     }
 
-    NSUInteger batchCount = 0, index = state->state, count = state->extra[1];
-
-    Class accessorClass = _objectSchema.accessorClass;
-    while (index < count && batchCount < len) {
-        // get acessor fot the object class
-        RLMObject *accessor = [[accessorClass alloc] initWithRealm:_realm schema:_objectSchema];
-        accessor->_row = (*_objectSchema.table)[[self indexInSource:index++]];
-        items->array[batchCount] = accessor;
-        buffer[batchCount] = accessor;
-        batchCount++;
-    }
-
-    for (NSUInteger i = batchCount; i < len; ++i) {
-        items->array[i] = nil;
-    }
-
-    state->itemsPtr = buffer;
-    state->state = index;
-    state->mutationsPtr = state->extra+1;
-
-    return batchCount;
+    return [enumerator countByEnumeratingWithState:state count:len];
 }
 
 - (NSUInteger)indexOfObjectWhere:(NSString *)predicateFormat, ... {
@@ -253,18 +324,12 @@ static inline void RLMResultsValidateInWriteTransaction(__unsafe_unretained RLMR
 
 - (id)valueForKey:(NSString *)key {
     RLMResultsValidate(self);
-    const size_t size = _backingView.size();
-    return RLMCollectionValueForKey(key, _realm, _objectSchema, size, ^size_t(size_t index) {
-        return _backingView.get_source_ndx(index);
-    });
+    return RLMCollectionValueForKey(self, key);
 }
 
 - (void)setValue:(id)value forKey:(NSString *)key {
     RLMResultsValidateInWriteTransaction(self);
-    const size_t size = _backingView.size();
-    RLMCollectionSetValueForKey(value, key, _realm, _objectSchema, size, ^size_t(size_t index) {
-        return _backingView.get_source_ndx(index);
-    });
+    RLMCollectionSetValueForKey(self, key, value);
 }
 
 - (RLMResults *)objectsWhere:(NSString *)predicateFormat, ... {
@@ -477,6 +542,12 @@ static NSNumber *averageOfProperty(TableType const& table, RLMRealm *realm, NSSt
     return _backingView.get_source_ndx(index);
 }
 
+- (realm::TableView)tableView {
+    RLMResultsValidateAttached(self);
+    // note: deliberately copies it
+    return _backingView;
+}
+
 @end
 
 @implementation RLMTableResults {
@@ -494,22 +565,6 @@ static NSNumber *averageOfProperty(TableType const& table, RLMRealm *realm, NSSt
 - (NSUInteger)count {
     RLMCheckThread(_realm);
     return _table->size();
-}
-
-- (id)valueForKey:(NSString *)key {
-    RLMResultsValidate(self);
-    const size_t size = _table->size();
-    return RLMCollectionValueForKey(key, _realm, _objectSchema, size, ^size_t(size_t index) {
-        return index;
-    });
-}
-
-- (void)setValue:(id)value forKey:(NSString *)key {
-    RLMResultsValidateInWriteTransaction(self);
-    const size_t size = _table->size();
-    RLMCollectionSetValueForKey(value, key, _realm, _objectSchema, size, ^size_t(size_t index) {
-        return index;
-    });
 }
 
 - (NSUInteger)indexOfObject:(RLMObject *)object {
@@ -560,7 +615,7 @@ static NSNumber *averageOfProperty(TableType const& table, RLMRealm *realm, NSSt
 
 - (void)deleteObjectsFromRealm {
     RLMResultsValidateInWriteTransaction(self);
-    RLMClearTable(_objectSchema);
+    RLMClearTable(self.objectSchema);
 }
 
 - (std::unique_ptr<Query>)cloneQuery {
@@ -569,6 +624,10 @@ static NSNumber *averageOfProperty(TableType const& table, RLMRealm *realm, NSSt
 
 - (NSUInteger)indexInSource:(NSUInteger)index {
     return index;
+}
+
+- (realm::TableView)tableView {
+    return _table->where().find_all();
 }
 @end