+++ Realm/RLMUtil.mm
@@ -188,16 +188,19 @@ NSDictionary *RLMDefaultValuesForObjectSchema(__unsafe_unretained RLMObjectSchem
     return defaults;
 }
 
-NSArray *RLMCollectionValueForKey(NSString *key, RLMRealm *realm, RLMObjectSchema *objectSchema, size_t count, size_t (^indexGenerator)(size_t)) {
+NSArray *RLMCollectionValueForKey(id<RLMFastEnumerable> collection, NSString *key) {
+    size_t count = collection.count;
     if (count == 0) {
         return @[];
     }
 
-    NSMutableArray *results = [NSMutableArray arrayWithCapacity:count];
+    RLMRealm *realm = collection.realm;
+    RLMObjectSchema *objectSchema = collection.objectSchema;
 
+    NSMutableArray *results = [NSMutableArray arrayWithCapacity:count];
     if ([key isEqualToString:@"self"]) {
         for (size_t i = 0; i < count; i++) {
-            size_t rowIndex = indexGenerator(i);
+            size_t rowIndex = [collection indexInSource:i];
             [results addObject:RLMCreateObjectAccessor(realm, objectSchema, rowIndex) ?: NSNull.null];
         }
         return results;
@@ -206,7 +209,7 @@ NSArray *RLMCollectionValueForKey(NSString *key, RLMRealm *realm, RLMObjectSchem
     RLMObjectBase *accessor = [[objectSchema.accessorClass alloc] initWithRealm:realm schema:objectSchema];
     realm::Table *table = objectSchema.table;
     for (size_t i = 0; i < count; i++) {
-        size_t rowIndex = indexGenerator(i);
+        size_t rowIndex = [collection indexInSource:i];
         accessor->_row = (*table)[rowIndex];
         RLMInitializeSwiftListAccessor(accessor);
         [results addObject:[accessor valueForKey:key] ?: NSNull.null];
@@ -215,14 +218,18 @@ NSArray *RLMCollectionValueForKey(NSString *key, RLMRealm *realm, RLMObjectSchem
     return results;
 }
 
-void RLMCollectionSetValueForKey(id value, NSString *key, RLMRealm *realm, RLMObjectSchema *objectSchema, size_t count, size_t (^indexGenerator)(size_t)) {
+void RLMCollectionSetValueForKey(id<RLMFastEnumerable> collection, NSString *key, id value) {
+    size_t count = collection.count;
     if (count == 0) {
         return;
     }
+
+    RLMRealm *realm = collection.realm;
+    RLMObjectSchema *objectSchema = collection.objectSchema;
     RLMObjectBase *accessor = [[objectSchema.accessorClass alloc] initWithRealm:realm schema:objectSchema];
     realm::Table *table = objectSchema.table;
     for (size_t i = 0; i < count; i++) {
-        size_t rowIndex = indexGenerator(i);
+        size_t rowIndex = [collection indexInSource:i];
         accessor->_row = (*table)[rowIndex];
         RLMInitializeSwiftListAccessor(accessor);
         [accessor setValue:value forKey:key];